+++ CHANGELOG.md
@@ -11,6 +11,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - You can now import [yal.cc](https://yal.cc/tools/pixel-font/) fonts directly from the json + png / aseprite file without having to go via ttf.
 - Added `set_palette` on `Object` and `ObjectAffine` to change the palette used by the sprite, this uses the same sprite data in vram just swapping out the palette. `SpriteVram` has a similar `with_palette` method.
+- Added palette manipulation to `GraphicsFrame` so you can synchronize palette swaps with background swaps.
+
+### Changed
+
+- Moved background colour DMA control from `VRAM_MANAGER` to `GraphicsFrame`.
 
 ## [0.23.1] - 2026/02/07